feat: updated some structure + moved validation logic into class
